Tidy up ABC
Removed always-open file handler as it was not really necessary. Also changed some exceptions to use NotImplementedErrors instead to clarify that the input was fine but that the mode itself was actually still unsupported.

Also renamed the get_node and get_element methods to comply with the previous version's interface - no reason to break compatibility here.