Move negative ID check to writer
The current implementation of the automatically assigned node IDs relies on the entity existing with a negative ID for a brief moment.

Adding checks to allow failing early on this check is more hassle than it's worth, so instead we just yell when we encounter the ID upon writing. This also checks that the user did not manually set a negative ID later.